+++ CHANGELOG.md
@@ -20,6 +20,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Refactor `enablePrefixer` to `addPrefixSelector`.
 
+### Fixed
+
+- Dynamically enable and disable the prefix selector setting when the feature is toggled.
+
 ## [0.2.2] - 2023-07-01
 
 ### Added